Add SetupObservability subcommand
