+++ zk_toolbox/crates/zk_inception/src/commands/ecosystem/mod.rs
@@ -10,6 +10,7 @@ mod change_default;
 mod create;
 pub mod create_configs;
 pub(crate) mod init;
+mod setup_observability;
 
 #[derive(Subcommand, Debug)]
 #[allow(clippy::large_enum_variant)]
@@ -23,6 +24,10 @@ pub enum EcosystemCommands {
     /// Change the default chain
     #[command(alias = "cd")]
     ChangeDefaultChain(ChangeDefaultChain),
+    /// Setup observability for the ecosystem,
+    /// downloading Grafana dashboards from the era-observability repo
+    #[command(alias = "obs")]
+    SetupObservability,
 }
 
 pub(crate) async fn run(shell: &Shell, args: EcosystemCommands) -> anyhow::Result<()> {
@@ -30,5 +35,6 @@ pub(crate) async fn run(shell: &Shell, args: EcosystemCommands) -> anyhow::Resul
         EcosystemCommands::Create(args) => create::run(args, shell),
         EcosystemCommands::Init(args) => init::run(args, shell).await,
         EcosystemCommands::ChangeDefaultChain(args) => change_default::run(args, shell),
+        EcosystemCommands::SetupObservability => setup_observability::run(shell),
     }
 }
